Add an spec for web/api/v1/products/:id
The branch for api scope had no coverage, so I'm adding a
basic one.

Note that is not a complete test suite for the endpoint since
we mean to have more comprenhensive ones inside the gem we
are 'importing': `Business::Api`, meaning the maintainer
should check the test suite located on the api folder.