Introducing Business::Schemas::Product
Which is similar to `Business::Schemas::BaseModel` we are just
changing the name in order to make it more accessible to our
consumers.

Since we Product is based on BaseModel it looks like a good idea
to introduce a shared example for both of them, so we don't
repeat ourselves.